+++ com.unity.render-pipelines.universal/CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [12.1.5] - 2022-01-14
 
+### Added
+- Added support for user-selected upscaling filters. Current options are automatic, bilinear, and nearest-neighbor.
+
+### Fixed
+- Fixed FXAA quality issues when render scale is not 1.0.
+
 Version Updated
 The version number for this package has increased due to a version update of a related graphics package.